houduan/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
